Fix #10999: Add support for materialized views in oracle
